+++ scripts/macros/homebrew/bg3/bg3.js
@@ -179,11 +179,10 @@ async function addFeatures(item, updates, options, id) {
     if (item.type != 'weapon') return;
     let actor = item.actor;
     if (!actor) return;
-    let proficient = item.system.proficient;
+    let proficient = item.system.proficient || item.actor.system.traits.weaponProf.value.has(baseItem);
     let baseItem = item.system.baseItem;
     if (!baseItem) return;
-    if (!proficient && isNewerVersion(game.version, '11.293')) {
-        proficient = item.actor.system.traits.weaponProf.value.has(baseItem);
+    if (!proficient) {
         if (simpleWeapons.includes(baseItem) && item.actor.system.traits.weaponProf.value.has('sim')) proficient = true;
         if (martialWeapons.includes(baseItem) && item.actor.system.traits.weaponProf.value.has('mar')) proficient = true;
     }