+++ scripts/macros/spells/compelledDuel.js
@@ -94,7 +94,7 @@ async function movement(token, updates, diff, id) {
         }
     };
     let oldDistance = chris.getDistance(sourceToken, fakeTargetToken);
-    await token.object._animation;
+    await token.object?._animation;
     let distance = chris.getDistance(sourceToken, token);
     if (oldDistance >= distance || distance <= 30) return;
     let turnCheck = chris.perTurnCheck(effect, 'spell', 'compelledDuel');